# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DocSecur API",
        "description": "API REST pour la plateforme de gestion médicale DocSecur. Authentification via Bearer Token (Laravel Passport).",
        "contact": {
            "email": "support@docsecur.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/admin/centres-sante": {
            "get": {
                "tags": [
                    "Admin - Centres de santé"
                ],
                "summary": "Lister les centres de santé",
                "description": "Retourne la liste paginée des centres de santé avec le nombre de médecins. Rôle requis : administrateur.",
                "operationId": "b3e1e826c0826f746f32f871afa1e169",
                "responses": {
                    "200": {
                        "description": "Liste des centres de santé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Centres de santé"
                ],
                "summary": "Créer un centre de santé",
                "description": "Ajoute un nouveau centre de santé. Rôle requis : administrateur.",
                "operationId": "33f69bd469be8c7eb5a9ec817e699adf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "adresse",
                                    "type",
                                    "region"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Hôpital Principal de Dakar"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Avenue Cheikh Anta Diop, Dakar"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "hopital",
                                            "clinique",
                                            "centre_sante",
                                            "poste_sante"
                                        ],
                                        "example": "hopital"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221338234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@hpd.sn"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Centre créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Centre de santé créé"
                                        },
                                        "centre": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/centres-sante/{id}": {
            "get": {
                "tags": [
                    "Admin - Centres de santé"
                ],
                "summary": "Détails d'un centre de santé",
                "operationId": "353baf1b73db4b12e35837390ef267f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du centre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centre non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Centres de santé"
                ],
                "summary": "Modifier un centre de santé",
                "operationId": "b84dead383a5852793dfc5d0af45398c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "hopital",
                                            "clinique",
                                            "centre_sante",
                                            "poste_sante"
                                        ]
                                    },
                                    "region": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centre modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Centre modifié"
                                        },
                                        "centre": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centre non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Centres de santé"
                ],
                "summary": "Supprimer un centre de santé",
                "operationId": "08fe94f3cbd2f7d9e65e12afa1805795",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centre supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Centre supprimé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centre non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}